feat: fetch balances in parallel
